+++ app/model/ExchangeRatesResponse.scala
@@ -7,7 +7,7 @@ import play.api.libs.json.{Format, Json}
 /**
   * Represents the response payload from the Exchange Rates API
   */
-case class ExchangeRatesResponse(base: CurrencyUnit, rates: Map[CurrencyUnit, Double])
+case class ExchangeRatesResponse(base: CurrencyUnit, rates: Map[CurrencyUnit, BigDecimal])
 
 object ExchangeRatesResponse {
   implicit val exchangeRatesResponseFormat: Format[ExchangeRatesResponse] = Json.format[ExchangeRatesResponse]
